Keep the delivery API outside the parent_controller chain
ContentController inherited RivetCms::ApplicationController, so a host
that re-parents the admin onto its own base class (config.parent_controller)
would also run that class's before_actions, layouts, and rescue_from
handlers on the token-gated JSON API, and the engine could only skip the
filters it knows about. The API now rides its own DeliveryBaseController
directly under ActionController::Base; the host-based org lookup both
sides shared moved to a concern. A structural spec pins the boundary.
